// frontend/src/services/OpenApi/models/OpeningDetailsOverviewDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { OpeningDetailsOverviewMilestoneDto } from './OpeningDetailsOverviewMilestoneDto';
import type { OpeningDetailsOverviewOpeningDto } from './OpeningDetailsOverviewOpeningDto';
export type OpeningDetailsOverviewDto = {
    opening: OpeningDetailsOverviewOpeningDto;
    milestones: OpeningDetailsOverviewMilestoneDto;
};